fix merge

+++ gframe/core_utils.cpp
@@ -336,7 +336,7 @@ CoreUtils::PacketStream ParseMessages(OCG_Duel duel) {
 void QueryStream::Parse(char*& buff) {
 	uint32_t size = BufferIO::Read<uint32_t>(buff);
 	char* current = buff;
-	while((current - buff) < size) {
+	while((uint32_t)(current - buff) < size) {
 		queries.emplace_back(current);
 	}
 }
